requirements.txt
kagglehub==0.2.7
tensorflow-cpu==2.14.0 